+++ storybook/src/stories/radio/Radio.template.tsx
@@ -1,5 +1,0 @@
-import { Radio, RadioProps } from '@equinor/fusion-react-radio';
-
-export const RadioTemplate = (props: RadioProps): JSX.Element => <Radio {...props} />;
-
-export default RadioTemplate;