+++ packages/filterpane/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filterpane@1.1.2...@equinor/fusion-react-filterpane@1.1.3) (2021-09-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-filterpane
+
+
+
+
+
 ## [1.1.2](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filterpane@1.1.1...@equinor/fusion-react-filterpane@1.1.2) (2021-09-16)